+++ EssentialFeed2Tests/Feed Cache/Helpers/FeedImageDataStoreSpy.swift
@@ -1,0 +1,37 @@
+//
+//  FeedImageDataStoreSpy.swift
+//  EssentialFeed2Tests
+//
+//  Created by mbp01 on 2021. 09. 27..
+//
+
+import Foundation
+import EssentialFeed2
+
+class FeedImageDataStoreSpy: FeedImageDataStore {
+
+    enum Messages: Equatable {
+        case insert(data: Data, for: URL)
+        case retrieve(dataFor: URL)
+    }
+
+    private(set) var receivedMessages = [Messages]()
+    private var retrievalCompletions = [(FeedImageDataStore.RetrievalResult) -> Void]()
+
+    func insert(data: Data, for url: URL, completion: @escaping (InsertionResult) -> Void) {
+        receivedMessages.append(.insert(data: data, for: url))
+    }
+
+    func retrieve(dataForURL url: URL, completion: @escaping (FeedImageDataStore.RetrievalResult) -> Void) {
+        receivedMessages.append(.retrieve(dataFor: url))
+        retrievalCompletions.append(completion)
+    }
+
+    func completeRetrieval(with error: Error, at index: Int = 0) {
+        retrievalCompletions[index](.failure(error))
+    }
+
+    func completeRetrieval(with data: Data?, at index: Int = 0) {
+        retrievalCompletions[index](.success(data))
+    }
+}
